+++ api/src/main/java/inventory/common/exception/CustomException.java
@@ -1,4 +1,4 @@
-package inventory.exception;
+package inventory.common.exception;
 
 import lombok.Getter;
