+++ api/src/main/java/inventory/common/exception/ExceptionCode.java
@@ -1,4 +1,4 @@
-package inventory.exception;
+package inventory.common.exception;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
