+++ api/src/main/java/inventory/common/exception/GlobalExceptionHandler.java
@@ -1,6 +1,6 @@
-package inventory.exception;
+package inventory.common.exception;
 
-import inventory.response.ApiResponse;
+import inventory.common.dto.response.ApiResponse;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
 import org.springframework.http.converter.HttpMessageNotReadableException;
